models/UTK/requirement.txt
pytorch == 1.10.2
pandas == 1.3.5
numpy == 1.21.6
scipy == 1.7.3